pyproject.toml
[project]
name = "fish-scale-analysis"
version = "0.1.0"
description = "Extract tubercle metrics from fish scale SEM images for species identification"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-image>=0.21.0",
    "tifffile>=2023.7.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "rich>=13.0.0",
    "rich-argparse>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
fish-scale-measure = "fish_scale_analysis.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fish_scale_analysis"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"
